requirements.txt
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
ffmpeg-python==0.2.0
Flask==1.1.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
future==0.18.2
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-datastore==1.13.2
google-cloud-iam==2.0.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
httplib2==0.18.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
libcst==0.3.10
MarkupSafe==1.1.1
matplotlib==3.2.2
mypy-extensions==0.4.3
numpy==1.18.1
opencv-python==4.2.0.34
proto-plus==1.7.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
SQLAlchemy==1.3.18
threadpoolctl==2.1.0
toml==0.10.1
typing-extensions==3.7.4.2
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
gunicorn 
